Move ScopeProperty functions in to scope.js
