+++ src/expression/node/FunctionNode.js
@@ -1,7 +1,8 @@
 import { isAccessorNode, isFunctionAssignmentNode, isIndexNode, isNode, isSymbolNode } from '../../utils/is.js'
 import { escape } from '../../utils/string.js'
 import { hasOwnProperty } from '../../utils/object.js'
-import { getSafeProperty, getScopeProperty, hasScopeProperty, validateSafeMethod } from '../../utils/customs.js'
+import { getSafeProperty, validateSafeMethod } from '../../utils/customs.js'
+import { getScopeProperty, hasScopeProperty } from '../../utils/scope.js'
 import { createSubScope } from '../../utils/scope.js'
 import { factory } from '../../utils/factory.js'
 import { defaultTemplate, latexFunctions } from '../../utils/latex.js'
