+++ src/expression/node/SymbolNode.js
@@ -1,5 +1,6 @@
 import { escape } from '../../utils/string.js'
-import { getSafeProperty, getScopeProperty, hasScopeProperty } from '../../utils/customs.js'
+import { getSafeProperty } from '../../utils/customs.js'
+import { getScopeProperty, hasScopeProperty } from '../../utils/scope.js'
 import { factory } from '../../utils/factory.js'
 import { toSymbol } from '../../utils/latex.js'
